requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# Document parsing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Audio processing & MP3 encoding
soundfile>=0.12.1
lameenc>=1.7.0

# Voice transcription (for auto-fill reference text)
openai-whisper>=20231117

# F5-TTS (voice cloning) — torch installed separately in run.bat for GPU/CPU detection
f5-tts>=1.1.0
